Dark theme and adjustments for print
The colors have all been turned into custom properties, which change depending on the theme. As for the print adjustments, i removed the "screen" part of some media queries that I tend to add by force of habit, because without them, the grid gets funky. I imagine this site is the kind of site someone might actually want to print, given it's TTRPG related, so fixes like this do matter.